Notification_with_colors and non-function search bar and sort by box. > > Co-authored-by: MirandaJaramillo mirandajaramillomorales@gmail.com
