+++ client/styles/style.css
@@ -1,0 +1,4 @@
+.gaugeChartClass {
+  margin: "-200px 0 -10px 0";
+  background-color: blue;
+}
