+++ package.json
@@ -27,6 +27,11 @@
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.9.0",
     "@mui/x-data-grid": "^5.11.0",
+    "@visx/group": "^2.10.0",
+    "@visx/heatmap": "^2.10.0",
+    "@visx/mock-data": "^2.1.2",
+    "@visx/responsive": "^2.10.0",
+    "@visx/scale": "^2.2.2",
     "chart.js": "^3.8.0",
     "chartjs-plugin-datalabels": "^2.0.0",
     "concurrently": "^7.2.2",
